1.7a complete
Added a game to show age and a game to answer love with percentages. In the game with age I added a "seconds" table and coded it to show how many seconds you are alive like the other times. In the love game I added alerts and a background change using javascript when different actions were done.